Move structs under data
